+++ src/index.ts
@@ -1,15 +1,19 @@
 /**
  * Welcome to namefully!
  *
- * `namefully` is a JavaScript utility for handing person names.
+ * `namefully` is a JavaScript utility for handling personal names.
  *
  * Sources
  * - repo: https://github.com/ralflorent/namefully
  * - docs: https://namefully.netlify.app
  * - npm: https://npmjs.com/package/namefully
+ * - jsr: https://jsr.io/@ralflorent/namefully
  *
  * @license MIT
  */
+import namefully from './namefully';
+
+export * from './builder';
 export * from './config';
 export { VERSION as version } from './constants';
 export * from './error';
@@ -19,3 +23,5 @@ export * from './namefully';
 export { Parser } from './parser';
 export * from './types';
 export { NameIndex } from './utils';
+
+export default namefully;